Lab 2 Traffic Lights
Pedestrian traffic lights allow any person who 
wishes to cross the road to press a button which triggers an 
alert for the trafic light system to change light into “green” 
and allow him to cross. You are asked to simulate a similar 
system using your arduino kit, 2 LEDs, a push button or a 
switch. Normal operation is that only one LED is active (ON) 
at any moment. A LED stays on for 2 seconds then it becomes 
(OFF) and the other LED becomes ON and stays for 2 
seconds then the cycle continues. If the switch or push button 
are pressed then the LEDs should invert their states instantly 
(The one that was (ON) becomes (OFF) and the one that was 
(OFF) becomes (ON)) and stay at these new states for 2 
seconds and then invert states for 2 seconds and so on. Every 
time the switch or push button are pressed the LED states 
must change instantly and the 2 seconds counter starts. In 
case of push button, you press and release . In case of switch, 
each press changes the state of the switch itself. NOTES: You 
should consider using resistors and apply the debouncing 
code.